+++ docs/changelog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 * Obsolete test data
+* Obsolete documentation figures
 
 ## [2.5.0] - 2023-09-05